// src/app/core/models/models.ts
// Function Data

export interface SignInData {
    nickname:                  string;
    password:               string;
    userType?:              string;
}

// State Data

export interface AuthData {
    accessToken:    string;
    client:         string;
    expiry:         string;
    tokenType:      string;
    uid:            string;
}

export interface UserData {
    id:             number;
    provider:       string;
    uid:            string;
    name:           string;
    nickname:       string;
    image:          any;
    email:          string;
}

// Configuration Options

export interface UserType {
    name:           string;
    path:           string;
}

export interface GlobalOptions {
    headers?:       { [key:string]: string; }
}

export interface TokenOptions {
    apiBase?:                   string;
    apiPath?:                   string;

    signInPath?:                string;
    signInRedirect?:            string;
    signInStoredUrlStorageKey?: string;

    signOutPath?:               string;
    validateTokenPath?:         string;
    signOutFailedValidate?:     boolean;

    deleteAccountPath?:         string;
    registerAccountPath?:       string;
    registerAccountCallback?:   string;

    updatePasswordPath?:        string;

    resetPasswordPath?:         string;
    resetPasswordCallback?:     string;

    userTypes?:                 UserType[];

    oAuthBase?:                 string;
    oAuthPaths?:                { [key:string]: string; };
    oAuthCallbackPath?:         string;
    oAuthWindowType?:           string;
    oAuthWindowOptions?:        { [key:string]: string; };

    globalOptions?:             GlobalOptions;
}